ti: soc: fix the poewr init issue when CONFIG_PM is disabled
Always compile power.c to ensure that TI's Power_init() is
unconditionally executed no matter the CONFIG_PM is enabled or
not.

The code implementing TI's power management required when enabling
CONFIG_PM is already conditionally included with the "#ifdef
CONFIG_PM" in power.c module. There is no need to do it again in the
CMakeLists.txt file.

Signed-off-by: Min Xu <m-xu@ti.com>