drivers: dma: cc23x0: Add power management
Add PM support to cc23x0 DMA module.

We assume that DMA clients (peripheral drivers or applications)
should take care of PM lock/unlock (pm_policy_state_lock_get/put).
This assumption is made for that SoC because:
- If a peripheral channel is used, then the transfer completion is
signaled on the peripheral's interrupt (handled in the DMA client
driver). This operating mode is specific to this SoC.
- If a software channel is used (memory-to-memory transfer), then
the transfer completion can be signaled to the application through
a callback.
Thus, in both cases, the PM can be unlocked at the right time by the
DMA client.

Signed-off-by: Julien Panis <jpanis@baylibre.com>