boards: ti: lp_em_cc2340r5x: enable onchip/offchip BLE FOTA on cc2340r5x
BLE FOTA and UART DFU was not supported on cc2340r5x devices because
MCUboot partitions were not defined in the devicetree. This update adds
the necessary devicetree nodes to define MCUboot partitions for onchip
operation. By default, MCUboot will use the internal flash (onchip) as
the secondary slot, but a conf and an overlay file can be added to the
build to switch to using a portion of the external SPI NOR (offchip)
as the secondary slot. In this update, snippets are also provided to
easily apply these conf and overlay files to a build.

Signed-off-by: Ging Gonzalez <g-gonzalez@ti.com>