boards: ti: lp_em_cc2340r53: use bt rx workq when mesh is enabled
In ble mesh, the system workq gets filled faster than it is freed,
causing excessive packet drops. This workq doesn't have enough
priority. In this commit, we now use the bt workq, which has higher
priority of -8. The tradeoff is that it adds RAM consumption as we
have to allocate a stack of size 2600bytes.

Signed-off-by: Ging Gonzalez <g-gonzalez@ti.com>