samples: bluetooth: fixing mesh example gen onoff client
The current offering for the BLE Mesh sample project does not properly support a generic onoff client. The sample claims to support one, however, the actual functionality is limited and a user is unable to set a publication address for the onoff client. The generic onoff client was fixed after defining a publication structure and altering the sent out mesh message.

Signed-off-by: Cameron Soviero <cam@ti.com>